pyproject.toml
[tool.black]
line-length = 100

[tool.poetry]
name = "sqlalchemy_multi_session_sample"
version = "0.1.0"
description = "sqlalchemy_multi_session_sample"
authors = ["tasogarei <tasogarei34@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
sqlalchemy = "^1.3.12"
mysqlclient = "^1.4.6"
python-dotenv = "^0.10.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
autopep8 = "^1.4.4"
black = "^19.10b0"
mypy = "^0.761"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
